moved the listing into a seperate component
usually we would use redux to fetch state values but they are being passed as props in this case. we should avoid passing too many props, as it may be need further by child components. we should avoid prop drilling.
here only single component is using those props so no issue here as it is a small app.
also if array built in methods are working no need to use another library